pkg/transform: rename to metricfamily, separate source files
This is just a mechanical refactor for better readability.
