+++ .gitignore
@@ -2,3 +2,4 @@
 /telemeter-server
 /authorization-server
 /_output
+telemeter-bench
