+++ pkg/http/server/diskstore.go
@@ -13,8 +13,8 @@ import (
 
 	clientmodel "github.com/prometheus/client_model/go"
 
+	"github.com/openshift/telemeter/pkg/metricfamily"
 	"github.com/openshift/telemeter/pkg/metricsclient"
-	"github.com/openshift/telemeter/pkg/transform"
 )
 
 type DiskStore struct {
@@ -124,7 +124,7 @@ func (s *DiskStore) WriteMetrics(ctx context.Context, partitionKey string, famil
 		return fmt.Errorf("unable to commit metrics to disk %s: %v", path, err)
 	}
 
-	metricSamples.WithLabelValues("disk").Add(float64(transform.Metrics(families)))
+	metricSamples.WithLabelValues("disk").Add(float64(metricfamily.MetricsCount(families)))
 
 	return nil
 }
