+++ pkg/metricfamily/anonymize.go
@@ -1,4 +1,4 @@
-package transform
+package metricfamily
 
 import (
 	"crypto/sha256"
