+++ pkg/metricfamily/none.go
@@ -1,0 +1,5 @@
+package metricfamily
+
+import clientmodel "github.com/prometheus/client_model/go"
+
+func None(*clientmodel.MetricFamily) (bool, error) { return true, nil }
